feat: eval from another .fnl file
- Moves to a single, named buffer for evaling awesome-fennel

This was a quick route to being able to find and use the process from
elsewhere, and might be fine going forward - seems like the use case
here is pretty specific to awesome, so maybe we only need one buffer
ever.

- two more functions for sending code to the repl

These need some work to handle cursor position in evil. Hoping to pull
from doom for that.

- add support for a few common globals (that were in my test case)

These could be pulled out and configurable, or we could just not
enforce global when compiling the fennel, as it is optional.

- require confirmation to kill the friar buffer

This keeps it around a bit longer for my setup, which seems to want to
kill the buffer too easily.

- some elisp formatting/indentation